Make 3d sfm example actually work.
The big fix is to swap the arguments to F_from_ransac(), see comment.

A corner in my appartment,
  Before this change: http://imgur.com/3DSZloa,dygIaCs#1
  After:              http://imgur.com/3DSZloa,dygIaCs#0
(The feature point that's in the corner in the "After" image belongs to
image pixels that are actually in the room's corner.)

Also remove a minus sign that caused the 3d data to be mirrored.