ch05ex02 (finding additional features on epipoles) mostly works
For every feature in image 1, this finds all features in image 2 that
are within a small distance of the first image's feature's epipolar
line, and if there's only one and the corresponding feature descriptors
are reasonably close, the two points are considered as matching points.

This doesn't particularly well (yet?), since it has a bunch of false
positives and since it's always possible to triangulate two matching 2d
points to a single 3d point as long as camera transforms are known,
there's no built-in protection for false matches.